Cleaned up FBX code Split code into reader and writer classes
