+++ dotnet/FbxImporter/Ara3D.FbxImporter.csproj
@@ -48,6 +48,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="FbxExporter.cs" />
     <Compile Include="FbxExtensions.cs" />
     <Compile Include="FbxImporter.cs" />
     <Compile Include="FbxImporterTests.cs" />
